backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pdfplumber
markdown
sentence-transformers
faiss-cpu
transformers
accelerate
openai